requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
openai==2.9.0
slack-sdk==3.27.1
python-telegram-bot==20.8
APScheduler==3.10.4
